Fix tmap bug (auto&). Able to solve own-strcmp-like function

+++ app/drtaint_marker/taint_map.h
@@ -2,15 +2,16 @@
 #define TAINT_MAP_H_
 
 #include "dr_api.h"
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-    bool tmap_has(instr_t *instr, byte taint);
+    bool tmap_has(instr_t *instr, uint32_t taint);
 
-    void tmap_emplace(instr_t *instr, byte taint);
+    void tmap_emplace(instr_t *instr, uint32_t taint);
 
     void tmap_print();
